Refactor: useStopAnimationOnResize 커스텀 훅 이름 변경
