+++ client/src/hooks/index.ts
@@ -1,5 +1,5 @@
 export { default as useRipple } from './use-ripple';
 export { default as useInput } from './use-input';
 export { default as useToggle } from './use-toggle';
-export { default as useStopAnimationOnResize } from './use-stop-animation-on-resize';
+export { default as useWindowResize } from './use-window-resize';
 export { default as useInterval } from './use-interval';
